+++ R/get_wasserportal_masters_data.R
@@ -1,8 +1,8 @@
 #' Wasserportal Berlin: get master data for a multiple stations
 #'
-#' @param master_urls urls with master data as retrieved by
-#'   \code{\link{get_stations}} and one of  "overview_list" sublist elements
-#'   column name "stammdaten_link"
+#' @param master_urls URLs to master data as found in column "stammdaten_link"
+#'   of the data frame returned by
+#'   \code{\link{get_stations}}\code{(type = "list")}
 #' @param run_parallel default: TRUE
 #'
 #' @return data frame with metadata for selected master urls
@@ -11,11 +11,12 @@
 #' @importFrom data.table rbindlist
 #' @examples
 #' \dontrun{
-#' stations <- wasserportal::get_stations()
-#' ### Reduce  to monitoring stations maintained by Berlin
-#' master_urls <- stations$overview_list$surface_water.water_level %>%
-#' dplyr::filter(.data$Betreiber == "Land Berlin") %>%
-#' dplyr::pull(.data$stammdaten_link)
+#' stations_list <- wasserportal::get_stations(type = "list")
+#'
+#' # Reduce  to monitoring stations maintained by Berlin
+#' master_urls <- stations_list$surface_water.water_level %>%
+#'   dplyr::filter(.data$Betreiber == "Land Berlin") %>%
+#'   dplyr::pull(.data$stammdaten_link)
 #'
 #' system.time(master_parallel <- get_wasserportal_masters_data(
 #'   master_urls
@@ -43,7 +44,7 @@ get_wasserportal_masters_data <- function(
     try(get_wasserportal_master_data(master_url))
   }
 
-  master_list <- kwb.utils::catAndRun(
+  master_list <- cat_and_run(
     messageText = sprintf(
       "Importing %d station metadata from Wasserportal Berlin",
       length(master_urls)
@@ -55,7 +56,7 @@ get_wasserportal_masters_data <- function(
     }
   )
 
-  failed <- sapply(master_list, kwb.utils::isTryError)
+  failed <- sapply(master_list, is_try_error)
 
   if (any(failed)) {
     message("Failed fetching data from the following URLs:")
@@ -71,24 +72,29 @@ get_wasserportal_masters_data <- function(
 #' \code{\link{get_wasserportal_stations_table}}
 #' @return data frame with metadata for selected station
 #' @importFrom dplyr mutate rename
-#' @importFrom kwb.utils stopFormatted
 #' @importFrom rlang .data
 #' @importFrom tidyr pivot_wider
 #' @export
 #' @examples
 #' \dontrun{
-#' stations <- wasserportal::get_stations()
+#' stations_list <- wasserportal::get_stations(type = "list")
 #'
-#' ## GW Station
-#' master_url <- stations$overview_list$groundwater.level$stammdaten_link[1]
+#' # GW Station
+#' master_url <- stations_list %>%
+#'   kwb.utils::selectElements("groundwater.level") %>%
+#'   kwb.utils::selectColumns("stammdaten_link")[1L]
+#'
 #' get_wasserportal_master_data(master_url)
 #'
-#' ## SW Station
-#' ### Reduce  to monitoring stations maintained by Berlin
-#' master_urls <- stations$overview_list$surface_water.water_level %>%
-#' dplyr::filter(.data$Betreiber == "Land Berlin") %>%
-#' dplyr::pull(.data$stammdaten_link)
-#' get_wasserportal_master_data(master_urls[1])
+#' # SW Station
+#'
+#' # Reduce  to monitoring stations maintained by Berlin
+#' master_urls <- stations_list %>%
+#'   kwb.utils::selectElements("surface_water.water_level") %>%
+#'   dplyr::filter(.data$Betreiber == "Land Berlin") %>%
+#'   dplyr::pull(.data$stammdaten_link)
+#'
+#' get_wasserportal_master_data(master_urls[1L])
 #' }
 #'
 get_wasserportal_master_data <- function(master_url)
@@ -101,13 +107,13 @@ get_wasserportal_master_data <- function(master_url)
     rvest::html_table()
 
   if (nrow(master_table) == 0L) {
-    kwb.utils::stopFormatted("No master table available at '%s'", master_url)
+    stop_formatted("No master table available at '%s'", master_url)
   }
 
   master_table %>%
     dplyr::rename("key" = "X1", "value" = "X2") %>%
     dplyr::mutate(key = stringr::str_remove_all(.data$key, "-")) %>%
-    dplyr::mutate(key = kwb.utils::substSpecialChars(.data$key)) %>%
+    dplyr::mutate(key = subst_special_chars(.data$key)) %>%
     tidyr::pivot_wider(names_from = "key", values_from = "value")
 }
 
@@ -116,7 +122,7 @@ stop_on_external_data_provider <- function(url)
 {
   if (is_external_link(url)) {
 
-    kwb.utils::stopFormatted(
+    stop_formatted(
       paste0(
         "The master_url '%s' you provided refers to an external ",
         "data provider. Currently only master data within '%s' can be ",