+++ R/get_wasserportal_stations.R
@@ -1,29 +1,28 @@
 # get_wasserportal_stations ----------------------------------------------------
 
 #' Get Names and IDs of the Stations of wasserportal.berlin.de
-#' 
+#'
 #' @param type one of "quality", "level", "flow"
 #' @export
 get_wasserportal_stations <- function(type = "quality")
 {
   if (! is.null(type)) {
-    type <- match.arg(type, c("quality", "level", "flow"))  
+    type <- match.arg(type, c("quality", "level", "flow"))
   }
-  
+
   file <- "stations_wasserportal.csv"
-  
+
   stations <- readPackageFile(file, fileEncoding = "UTF-8")
-  
-  get <- kwb.utils::selectColumns
-  
-  stations$id <- as.character(get(stations, "id"))
-  stations$name <- kwb.utils::substSpecialChars(get(stations, "name"))
-  
+
+  stations$id <- as.character(select_columns(stations, "id"))
+  stations$name <- subst_special_chars(select_columns(stations, "name"))
+
   is_available <- if (is.null(type)) {
     seq_len(nrow(stations))
   } else {
-    nzchar(get(stations, type))
+    nzchar(select_columns(stations, type))
   }
-  
-  kwb.utils::toLookupList(data = get(stations, c("name", "id"))[is_available, ])
+
+  select_columns(stations, c("name", "id"))[is_available, ] %>%
+    to_lookup_list(data = .)
 }