+++ NEWS.md
@@ -1,3 +1,12 @@
+# [wasserportal 0.4.0](https://github.com/KWB-R/wasserportal/releases/tag/v0.4.0) <small>2024-04-05</small>
+
+* New feature: add support for downloading all available surface water quality 
+data for one or multiple monitoring stations. For details see `get_surfacewater_qualities()` 
+* Bugfix for groundwater level and quality due to new Wasserportal API 
+* Add project [AD4GD](https://www.kompetenz-wasser.de/de/forschung/projekte/ad4gd) 
+as funder
+
+
 # [wasserportal 0.3.0](https://github.com/KWB-R/wasserportal/releases/tag/v0.3.0) <small>2023-02-19</small>
 
 * Fix errors in GitHub actions: use actions from branches `v2`, `v3`, not from